moved stuff to cogs/diepio
